requirements.txt
langgraph
langchain
langchain-google-genai
chromadb
networkx
pydantic
python-dotenv
langsmith
mcp
jupyter
ruff
pytest
